SpongeBob-Case Conversion for Data Frames

Convert data.frame objects to Mocking SpongeBob case. This is the S3 method for data.frame objects. Generally, you will not be using this method directly; instead, use the S3 generic tospongebob. If you really need to use this method, you will need to do spongebob:::tospongebob.data.frame because it is an internal method.

This method will convert character and factor columns to Mocking SpongeBob case. It will also optionlly convert column and row names, enabled by default.

# S3 method for data.frame
tospongebob(x, ..., convert.rownames = TRUE,
  convert.colnames = TRUE)

Arguments

x
data.frame with text to be converted to Mocking SpongeBob case.

...
onlY hERe tO sATisFy R CMd CHeCk.

convert.rownames
logical, indicating whether or not to convert the row names of the data.frame

convert.colnames
logical, indicating whether or not to convert the column names of the data.frame

Value

data.frame with its text converted to Mocking SpongeBob case.

See also

Examples

df1 <- head(datasets::CO2)
tospongebob(df1)

#>   PLAnt   tYPe  trEaTmENt cONc UPTakE
#> 1   qN1 QUeBEc nonCHilLed   95   16.0
#> 2   qN1 QUeBEc nonCHilLed  175   30.4
#> 3   qN1 QUeBEc nonCHilLed  250   34.8
#> 4   qN1 QUeBEc nonCHilLed  350   37.2
#> 5   qN1 QUeBEc nonCHilLed  500   35.3
#> 6   qN1 QUeBEc nonCHilLed  675   39.2

df2 <- head(datasets::mtcars)
tospongebob(df2)

#>                    mPg CYl Disp  hP dRaT    WT  QSEc vs Am geaR carB
#> mAZDa rX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> maZda Rx4 waG     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> daTSun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> HoRNeT 4 DrIVe    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> hOrnET SpOrtaBOUt 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> VAliANt           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1